Modified install process
install.php
- Added a $server_port variable/check to the BASE_URL constant
definition, for instances when installing Secretary on a port other than
80.

Added a default user display name of "Administrator" to the initial user
insert query, this was in response to errors encountered when the
setup.sql script was run complaining that the display_name field had no
default value.

setup.sql
- Removed NOT NULL from the definitions for display_name and level_id in
the users table, level_id does not appear to be used at the moment and
display name appears to be optional.